Finish unused config (barely)
