+++ package.json
@@ -23,13 +23,13 @@
   },
   "homepage": "https://github.com/danmindru/responsive-html-email-signature#readme",
   "scripts": {
-    "start": "gulp",
-    "once": "gulp run-pipeline",
+    "start": "node ./node_modules/.bin/gulp",
+    "once": "node ./node_modules/.bin/gulp run-pipeline",
     "deploy": "npm run test && cp -r dist demo && git push origin `git subtree split --prefix demo develop`:gh-pages --force",
     "test": "npm run once && node ./node_modules/.bin/ava",
     "test:watch": "npm run once && node ./node_modules/.bin/ava --watch",
-    "format": "prettier {tasks,tests}/**/*.js gulpfile.js .eslintrc.js --write",
-    "lint": "eslint ./**/*.js gulpfile.js"
+    "format": "node ./node_modules/.bin/prettier {tasks,tests}/**/*.js gulpfile.js .eslintrc.js --write",
+    "lint": "node ./node_modules/.bin/eslint ./**/*.js gulpfile.js"
   },
   "dependencies": {
     "autoprefixer": "^9.6.1",
@@ -75,12 +75,12 @@
   "husky": {
     "hooks": {
       "pre-push": "npm run test",
-      "pre-commit": "npm run lint && ./node_modules/.bin/pretty-quick --staged --pattern ./**/*.js"
+      "pre-commit": "npm run lint && node ./node_modules/.bin/pretty-quick --staged --pattern ./**/*.js"
     }
   },
   "ava": {
     "helpers": [
-			"**/util.js"
-		]
+      "**/util.js"
+    ]
   }
 }
