--- tests/unit/util.js
@@ -1,3 +1,0 @@
-const test = require('ava');
-
-test('noop', async t => { t.deepEqual('', ''); });
